% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.runner-package.R
\docType{package}
\name{pacta.runner-package}
\alias{pacta.runner}
\alias{pacta.runner-package}
\title{pacta.runner: Run the PACTA process with a single command}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alex Axthelm \email{AAxthelm@RMI.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID})

}
\keyword{internal}
